pyproject.toml
[project]
name = "PixPulse"
version = "0.1.0"
description = ""
authors = [
    {name = "Lincoln",email = "1155176490@link.cuhk.edu.hk"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
name = "PixPulse"
packages = [
    {include = "routers"},
    {include = "utils"},
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
